Move CallParameterSets to be a part of the WhereFilter protocol.
